Initial commit
Basic scripts to take a dump of a live db and place it in an existing
DB used for development. After the DB clone a search and replace is
carried out to replace the hostname of the Wordpress install.
Parameters that need setting are in the dblive2stage.php file.